pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pydantic-agent"
version = "0.1.0"
description = "Pydantic-based code assistant for VS Code"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=0.19.0",
    "async-timeout>=4.0.0",
    "asyncio>=3.4.3",
    "rich>=13.7.0",
    "prompt_toolkit>=3.0.43"
]
